test(e2e): target-gate the harness to Linux; search the window tree in depth
All three PR #715 review findings: x11rb moves to a
target_os = "linux" dependency with a stub main elsewhere, so a
Windows/macOS workspace build never compiles X11 code; the window
search walks the tree breadth-first to depth 4 (a WM reparents clients
under decoration frames — root's direct children only worked on bare
Xvfb) and accepts any WM_NAME property type, not just STRING.
